pr feedback refactor
