+++ src/routes/applications.ts
@@ -15,6 +15,7 @@ import {
   sendEmail,
   searchCollaboratorApplications,
   getApplicationUpdates,
+  createAppHistoryTSV,
 } from '../domain/service';
 import { BadRequest } from '../utils/errors';
 import logger from '../logger';
@@ -28,7 +29,7 @@ import { Transporter } from 'nodemailer';
 import SMTPTransport from 'nodemailer/lib/smtp-transport';
 // https://www.archiverjs.com/docs/quickstart
 import moment from 'moment';
-import { getSearchParams, createDacoCSVFile, encrypt, createAppHistoryTSV } from '../utils/misc';
+import { getSearchParams, createDacoCSVFile, encrypt } from '../utils/misc';
 import JSZip from 'jszip';
 
 export interface IRequest extends Request {
